:bento: (llm): custom lock screen banner
