+++ apps/ledger-live-mobile/src/screens/CustomImage/EuropaWelcomeView.tsx
@@ -1,0 +1,3 @@
+import StaxWelcomeView from "./StaxWelcomeView";
+
+export default StaxWelcomeView;
