+++ apps/ledger-live-mobile/src/screens/CustomImage/Step0Welcome.tsx
@@ -1,10 +1,8 @@
 import React, { useCallback, useState } from "react";
 import { StackScreenProps } from "@react-navigation/stack";
 import { SafeAreaView } from "react-native-safe-area-context";
-import Video from "react-native-video";
 import { Flex, Text } from "@ledgerhq/native-ui";
 import { useTranslation } from "react-i18next";
-import { useTheme } from "@react-navigation/native";
 import { getDeviceModel } from "@ledgerhq/devices";
 import CustomImageBottomModal from "~/components/CustomImage/CustomImageBottomModal";
 import BottomButtonsContainer from "~/components/CustomImage/BottomButtonsContainer";
@@ -12,14 +10,9 @@ import Button from "~/components/wrappedUi/Button";
 import { ScreenName } from "~/const";
 import { CustomImageNavigatorParamList } from "~/components/RootNavigator/types/CustomImageNavigator";
 import { TrackScreen } from "~/analytics";
-import videoSources from "../../../assets/videos";
-import { useSystem } from "~/hooks";
 import { DeviceModelId } from "@ledgerhq/types-devices";
-
-const videoDimensions = {
-  height: 550,
-  width: 1080,
-};
+import EuropaWelcomeView from "./EuropaWelcomeView";
+import StaxWelcomeView from "./StaxWelcomeView";
 
 const analyticsScreenName = "Introduction of the customization flow";
 const analyticsButtonEventProps = {
@@ -29,7 +22,6 @@ const analyticsButtonEventProps = {
 const Step0Welcome: React.FC<
   StackScreenProps<CustomImageNavigatorParamList, ScreenName.CustomImageStep0Welcome>
 > = ({ route }) => {
-  const theme = useTheme();
   const [modalOpened, setModalOpened] = useState(false);
   const { t } = useTranslation();
 
@@ -46,28 +38,13 @@ const Step0Welcome: React.FC<
     setModalOpened(false);
   }, [setModalOpened]);
 
-  const { screen } = useSystem();
-
   return (
     <SafeAreaView style={{ flex: 1 }} edges={["bottom"]}>
       <TrackScreen category={analyticsScreenName} />
       <Flex flex={1} mt={8} justifyContent={"space-between"}>
         <Flex>
-          <Video
-            disableFocus
-            source={
-              theme.dark
-                ? videoSources.customLockScreenBannerDark
-                : videoSources.customLockScreenBannerLight
-            }
-            style={{
-              width: screen.width,
-              height: (videoDimensions.height / videoDimensions.width) * screen.width,
-            }}
-            muted
-            repeat
-            resizeMode={"contain"}
-          />
+          {deviceModelId === DeviceModelId.europa ? <EuropaWelcomeView /> : <StaxWelcomeView />}
+
           <Flex px={7} mt={11}>
             <Text
               variant="h4"
